move implementations of AsyncOperationResult to internal, rename executor to performer
