+++ geode-management/src/test/java/org/apache/geode/management/api/ClientAsyncOperationResultTest.java
@@ -23,6 +23,8 @@ import java.util.concurrent.TimeoutException;
 
 import org.junit.Test;
 
+import org.apache.geode.management.internal.ClientAsyncOperationResult;
+
 public class ClientAsyncOperationResultTest {
 
   @Test
